requirements.txt
Flask>=2.2.5
Flask-Login>=0.6.2
Flask-WTF>=1.1.1
Flask-Migrate>=4.0.4
alembic>=1.11.1
Flask-Limiter>=3.0.2
Flask-SQLAlchemy>=3.0.3
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
argon2-cffi>=21.3.0
Pillow>=10.0.1
gunicorn>=20.1.0
